requirements.txt
tidy3d>=2.7
numpy>=1.24
scipy>=1.10
pandas>=2.0
matplotlib>=3.7
python-dotenv>=1.0
